Changed some file structure, made landing page the first page, some css for signup
