+++ client/src/components/Login.js
@@ -26,9 +26,9 @@ class Login extends Component {
     }
 
     componentWillMount() {
-        if (this.Auth.loggedIn()) {
-            this.props.history.replace('/');
-        }
+        // if (this.Auth.loggedIn()) {
+        //     this.props.history.replace('/');
+        // }
     }
 
     handleFormSubmit = event => {
