+++ client/src/components/Signup/Signup.css
@@ -1,0 +1,3 @@
+#logo {
+    width: 100%;
+}
