+++ client/src/components/Signup/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Signup";
